+++ src/constants.ts
@@ -10,7 +10,6 @@ export const commandMap = {
   PAUSE_SHIFT_INTERVAL: "shifty.pauseShiftInterval",
   RESET_SHIFT_INTERVAL: "shifty.resetShiftInterval",
   SHOW_STATUS: "shifty.showStatus",
-  ENABLE_DEBUGGING: "shifty.enableDebugging",
 };
 
 export const DEFAULT_COLOR_THEME = "Default Dark+";